+++ src/startup/genx_startup.jl
@@ -1,7 +1,3 @@
-function __init__()
-    print_genx_version()
-end
-
 function print_genx_version()
     v = pkgversion(GenX)
     ascii_art = raw"""
@@ -54,6 +50,9 @@ Returns `nothing`.
 """
 function _precompile()
     @info "Running precompile script for GenX. This may take a few minutes."
+    @info "If you want to skip this step, please set the environment variable " *
+        "`GENX_PRECOMPILE` to `false` before running `using GenX`. \n" *
+        "Example: `ENV[\"GENX_PRECOMPILE\"] = \"false\"`."
     redirect_stdout(devnull) do
         warnerror_logger = ConsoleLogger(stderr, Logging.Warn)
         with_logger(warnerror_logger) do