+++ src/write_outputs/energy_share_requirement/write_esr_revenue.jl
@@ -1,5 +1,5 @@
 @doc raw"""
-	write_esr_revenue(path::AbstractString, inputs::Dict, setup::Dict, dfPower::DataFrame, dfESR::DataFrame, EP::Model)
+    write_esr_revenue(path::AbstractString, inputs::Dict, setup::Dict, dfPower::DataFrame, dfESR::DataFrame, EP::Model)
 
 Function for reporting the renewable/clean credit revenue earned by each generator listed in the input file. GenX will print this file only when RPS/CES is modeled and the shadow price can be obtained form the solver. Each row corresponds to a generator, and each column starting from the 6th to the second last is the total revenue earned from each RPS constraint. The revenue is calculated as the total annual generation (if elgible for the corresponding constraint) multiplied by the RPS/CES price. The last column is the total revenue received from all constraint. The unit is \$.
 """
@@ -15,6 +15,7 @@ function write_esr_revenue(path::AbstractString,
     zones = zone_id.(gen)
     rid = resource_id.(gen)
 
+    FUSION = ids_with(gen, :fusion)
     dfESRRev = DataFrame(region = regions,
         Resource = inputs["RESOURCE_NAMES"],
         zone = zones,
@@ -39,44 +40,27 @@ function write_esr_revenue(path::AbstractString,
         esr_col = Symbol("ESR_$i")
         price = dfESR[i, :ESR_Price]
         derated_annual_net_generation = dfPower[1:G, :AnnualSum] .* esr.(gen, tag = i)
+        derated_annual_net_generation[FUSION] .+= thermal_fusion_annual_parasitic_power(
+            EP, inputs, setup) .* esr.(gen[FUSION], tag = i)
         revenue = derated_annual_net_generation * price
         dfESRRev[!, esr_col] = revenue
 
         if !isempty(VRE_STOR)
             if !isempty(SOLAR_ONLY)
-                solar_resources = ((gen_VRE_STOR.wind .== 0) .& (gen_VRE_STOR.solar .!= 0))
-                dfESRRev[SOLAR, esr_col] = (value.(EP[:vP_SOLAR][SOLAR, :]).data
-                                            .*
-                                            etainverter.(gen_VRE_STOR[solar_resources]) *
-                                            weight) .*
-                                           esr_vrestor.(gen_VRE_STOR[solar_resources],
-                    tag = i) * price
+                dfESRRev[SOLAR, esr_col] = (value.(EP[:vP_SOLAR][SOLAR, :]).data .*
+                                            etainverter.(gen[SOLAR]) * weight) .*
+                                            esr_vrestor.(gen[SOLAR], tag = i) * price
             end
             if !isempty(WIND_ONLY)
-                wind_resources = ((gen_VRE_STOR.wind .!= 0) .& (gen_VRE_STOR.solar .== 0))
-                dfESRRev[WIND, esr_col] = (value.(EP[:vP_WIND][WIND, :]).data
-                                           *
-                                           weight) .*
-                                          esr_vrestor.(gen_VRE_STOR[wind_resources],
-                    tag = i) * price
+                dfESRRev[WIND, esr_col] = (value.(EP[:vP_WIND][WIND, :]).data * weight) .*
+                                            esr_vrestor.(gen[WIND], tag = i) * price
             end
             if !isempty(SOLAR_WIND)
-                solar_and_wind_resources = ((gen_VRE_STOR.wind .!= 0) .&
-                                            (gen_VRE_STOR.solar .!= 0))
-                dfESRRev[SOLAR_WIND, esr_col] = (((value.(EP[:vP_WIND][SOLAR_WIND,
-                    :]).data * weight)
-                                                  .*
-                                                  esr_vrestor.(
-                    gen_VRE_STOR[solar_and_wind_resources],
-                    tag = i) * price) +
-                                                 (value.(EP[:vP_SOLAR][SOLAR_WIND, :]).data
-                                                  .*
-                                                  etainverter.(gen_VRE_STOR[solar_and_wind_resources])
-                                                  *
-                                                  weight) .*
-                                                 esr_vrestor.(
-                    gen_VRE_STOR[solar_and_wind_resources],
-                    tag = i) * price)
+                dfESRRev[SOLAR_WIND, esr_col] = (((value.(EP[:vP_WIND][SOLAR_WIND, :]).data * weight) .*
+                                                    esr_vrestor.(gen[SOLAR_WIND], tag = i) * price) +
+                                                    (value.(EP[:vP_SOLAR][SOLAR_WIND, :]).data .*
+                                                    etainverter.(gen[SOLAR_WIND]) * weight) .*
+                                                    esr_vrestor.(gen[SOLAR_WIND], tag = i) * price)
             end
         end
     end