+++ src/write_outputs/write_charge.jl
@@ -4,49 +4,53 @@
 Function for writing the charging energy values of the different storage technologies.
 """
 function write_charge(path::AbstractString, inputs::Dict, setup::Dict, EP::Model)
-    gen = inputs["RESOURCES"]
+    gen = inputs["RESOURCES"]   # Resources (objects) 
+    resources = inputs["RESOURCE_NAMES"]    # Resource names
     zones = zone_id.(gen)
 
-    G = inputs["G"]     # Number of resources (generators, storage, DR, and DERs)
     T = inputs["T"]     # Number of time steps (hours)
     STOR_ALL = inputs["STOR_ALL"]
     FLEX = inputs["FLEX"]
     ELECTROLYZER = inputs["ELECTROLYZER"]
     VRE_STOR = inputs["VRE_STOR"]
     VS_STOR = !isempty(VRE_STOR) ? inputs["VS_STOR"] : []
+    FUSION = ids_with(gen, :fusion)
 
-    # Power withdrawn to charge each resource in each time step
-    dfCharge = DataFrame(Resource = inputs["RESOURCE_NAMES"],
-        Zone = zones,
-        AnnualSum = Array{Union{Missing, Float64}}(undef, G))
-    charge = zeros(G, T)
-
+    weight = inputs["omega"]
     scale_factor = setup["ParameterScale"] == 1 ? ModelScalingFactor : 1
+
+    charge = Matrix[]
+    charge_ids = Vector{Int}[]
     if !isempty(STOR_ALL)
-        charge[STOR_ALL, :] = value.(EP[:vCHARGE][STOR_ALL, :]) * scale_factor
+        push!(charge, value.(EP[:vCHARGE]))
+        push!(charge_ids, STOR_ALL)
     end
     if !isempty(FLEX)
-        charge[FLEX, :] = value.(EP[:vCHARGE_FLEX][FLEX, :]) * scale_factor
+        push!(charge, value.(EP[:vCHARGE_FLEX]))
+        push!(charge_ids, FLEX)
     end
     if (setup["HydrogenMinimumProduction"] > 0) & (!isempty(ELECTROLYZER))
-        charge[ELECTROLYZER, :] = value.(EP[:vUSE][ELECTROLYZER, :]) * scale_factor
+        push!(charge, value.(EP[:vUSE]))
+        push!(charge_ids, ELECTROLYZER)
     end
     if !isempty(VS_STOR)
-        charge[VS_STOR, :] = value.(EP[:vCHARGE_VRE_STOR][VS_STOR, :]) * scale_factor
+        push!(charge, value.(EP[:vCHARGE_VRE_STOR]))
+        push!(charge_ids, VS_STOR)
     end
-
-    dfCharge.AnnualSum .= charge * inputs["omega"]
-
-    filepath = joinpath(path, "charge.csv")
-    if setup["WriteOutputs"] == "annual"
-        write_annual(filepath, dfCharge)
-    else # setup["WriteOutputs"] == "full"
-        df_Charge = write_fulltimeseries(filepath, charge, dfCharge)
-        if setup["OutputFullTimeSeries"] == 1 && setup["TimeDomainReduction"] == 1
-            write_full_time_series_reconstruction(
-                path, setup, df_Charge, "charge")
-            @info("Writing Full Time Series for Charge")
-        end
+    if !isempty(FUSION)
+        _, mat = prepare_fusion_parasitic_power(EP, inputs)
+        push!(charge, mat)
+        push!(charge_ids, FUSION)
     end
+    charge = reduce(vcat, charge, init = zeros(0, T))
+    charge_ids = reduce(vcat, charge_ids, init = Int[])
+
+    charge *= scale_factor
+
+    df = DataFrame(Resource = resources[charge_ids],
+        Zone = zones[charge_ids])
+    df.AnnualSum = charge * weight
+
+    write_temporal_data(df, charge, path, setup, "charge")
     return nothing
 end