+++ src/write_outputs/write_curtailment.jl
@@ -5,45 +5,42 @@ Function for writing the curtailment values of the different variable renewable 
 	co-located).
 """
 function write_curtailment(path::AbstractString, inputs::Dict, setup::Dict, EP::Model)
-    gen = inputs["RESOURCES"]
+    gen = inputs["RESOURCES"]  # Resources (objects)
+    resources = inputs["RESOURCE_NAMES"] # Resource names
+    zones = zone_id.(gen)
+
     G = inputs["G"]     # Number of resources (generators, storage, DR, and DERs)
     T = inputs["T"]     # Number of time steps (hours)
     VRE = inputs["VRE"]
-    dfCurtailment = DataFrame(Resource = inputs["RESOURCE_NAMES"],
-        Zone = zone_id.(gen),
-        AnnualSum = zeros(G))
-    curtailment = zeros(G, T)
+    VRE_STOR = inputs["VRE_STOR"]
+
+    weight = inputs["omega"]
     scale_factor = setup["ParameterScale"] == 1 ? ModelScalingFactor : 1
-    curtailment[VRE, :] = scale_factor *
-                          (value.(EP[:eTotalCap][VRE]) .* inputs["pP_Max"][VRE, :] .-
+
+    curtailment = zeros(G, T)
+    curtailment[VRE, :] = (value.(EP[:eTotalCap][VRE]) .* inputs["pP_Max"][VRE, :] .-
                            value.(EP[:vP][VRE, :]))
 
-    VRE_STOR = inputs["VRE_STOR"]
     if !isempty(VRE_STOR)
         SOLAR = setdiff(inputs["VS_SOLAR"], inputs["VS_WIND"])
         WIND = setdiff(inputs["VS_WIND"], inputs["VS_SOLAR"])
         SOLAR_WIND = intersect(inputs["VS_SOLAR"], inputs["VS_WIND"])
-        gen_VRE_STOR = gen.VreStorage
         if !isempty(SOLAR)
-            curtailment[SOLAR, :] = scale_factor *
-                                    (value.(EP[:eTotalCap_SOLAR][SOLAR]).data .*
-                                     inputs["pP_Max_Solar"][SOLAR, :] .-
-                                     value.(EP[:vP_SOLAR][SOLAR, :]).data) .*
-                                    etainverter.(gen_VRE_STOR[(gen_VRE_STOR.solar .!= 0)])
+            curtailment[SOLAR, :] = (value.(EP[:eTotalCap_SOLAR][SOLAR]).data .*
+                                    inputs["pP_Max_Solar"][SOLAR, :] .-
+                                    value.(EP[:vP_SOLAR][SOLAR, :]).data) .*
+                                    etainverter.(gen[SOLAR])
         end
         if !isempty(WIND)
-            curtailment[WIND, :] = scale_factor * (value.(EP[:eTotalCap_WIND][WIND]).data .*
+            curtailment[WIND, :] = (value.(EP[:eTotalCap_WIND][WIND]).data .*
                                     inputs["pP_Max_Wind"][WIND, :] .-
                                     value.(EP[:vP_WIND][WIND, :]).data)
         end
         if !isempty(SOLAR_WIND)
-            curtailment[SOLAR_WIND, :] = scale_factor *
-                                         ((value.(EP[:eTotalCap_SOLAR])[SOLAR_WIND].data
-                                           .*
+            curtailment[SOLAR_WIND, :] = ((value.(EP[:eTotalCap_SOLAR])[SOLAR_WIND].data .*
                                            inputs["pP_Max_Solar"][SOLAR_WIND, :] .-
-                                           value.(EP[:vP_SOLAR][SOLAR_WIND, :]).data)
-                                          .*
-                                          etainverter.(gen_VRE_STOR[((gen_VRE_STOR.wind .!= 0) .& (gen_VRE_STOR.solar .!= 0))])
+                                           value.(EP[:vP_SOLAR][SOLAR_WIND, :]).data) .*
+                                          etainverter.(gen[SOLAR_WIND])
                                           +
                                           (value.(EP[:eTotalCap_WIND][SOLAR_WIND]).data .*
                                            inputs["pP_Max_Wind"][SOLAR_WIND, :] .-
@@ -51,17 +48,13 @@ function write_curtailment(path::AbstractString, inputs::Dict, setup::Dict, EP::
         end
     end
 
-    dfCurtailment.AnnualSum = curtailment * inputs["omega"]
+    curtailment *= scale_factor
 
-    filename = joinpath(path, "curtail.csv")
-    if setup["WriteOutputs"] == "annual"
-        write_annual(filename, dfCurtailment)
-    else # setup["WriteOutputs"] == "full"
-        df_Curtailment = write_fulltimeseries(filename, curtailment, dfCurtailment)
-        if setup["OutputFullTimeSeries"] == 1 && setup["TimeDomainReduction"] == 1
-            write_full_time_series_reconstruction(path, setup, df_Curtailment, "curtail")
-            @info("Writing Full Time Series for Curtailment")
-        end
-    end
+    df = DataFrame(Resource = resources,
+        Zone = zones,
+        AnnualSum = zeros(G))
+    df.AnnualSum = curtailment * weight
+
+    write_temporal_data(df, curtailment, path, setup, "curtailment")
     return nothing
 end