+++ src/write_outputs/write_storage.jl
@@ -4,7 +4,8 @@
 Function for writing the capacities of different storage technologies, including hydro reservoir, flexible storage tech etc.
 """
 function write_storage(path::AbstractString, inputs::Dict, setup::Dict, EP::Model)
-    gen = inputs["RESOURCES"]
+    gen = inputs["RESOURCES"]   # Resources (objects)
+    resources = inputs["RESOURCE_NAMES"]   # Resource names
     zones = zone_id.(gen)
 
     T = inputs["T"]     # Number of time steps (hours)
@@ -15,33 +16,29 @@ function write_storage(path::AbstractString, inputs::Dict, setup::Dict, EP::Mode
     VRE_STOR = inputs["VRE_STOR"]
     VS_STOR = !isempty(VRE_STOR) ? inputs["VS_STOR"] : []
 
-    # Storage level (state of charge) of each resource in each time step
-    dfStorage = DataFrame(Resource = inputs["RESOURCE_NAMES"], Zone = zones)
-    storagevcapvalue = zeros(G, T)
+    weight = inputs["omega"]
+    scale_factor = setup["ParameterScale"] == 1 ? ModelScalingFactor : 1
 
-    if !isempty(inputs["STOR_ALL"])
-        storagevcapvalue[STOR_ALL, :] = value.(EP[:vS][STOR_ALL, :])
+    stored = Matrix[]
+    if !isempty(STOR_ALL)
+        push!(stored, value.(EP[:vS]))
     end
-    if !isempty(inputs["HYDRO_RES"])
-        storagevcapvalue[HYDRO_RES, :] = value.(EP[:vS_HYDRO][HYDRO_RES, :])
+    if !isempty(HYDRO_RES)
+        push!(stored, value.(EP[:vS_HYDRO]))
     end
-    if !isempty(inputs["FLEX"])
-        storagevcapvalue[FLEX, :] = value.(EP[:vS_FLEX][FLEX, :])
+    if !isempty(FLEX)
+        push!(stored, value.(EP[:vS_FLEX]))
     end
     if !isempty(VS_STOR)
-        storagevcapvalue[VS_STOR, :] = value.(EP[:vS_VRE_STOR][VS_STOR, :])
+        push!(stored, value.(EP[:vS_VRE_STOR]))
     end
-    if setup["ParameterScale"] == 1
-        storagevcapvalue *= ModelScalingFactor
-    end
+    stored = reduce(vcat, stored, init = zeros(0, T))
+    stored *= scale_factor
 
-    dfStorage = hcat(dfStorage, DataFrame(storagevcapvalue, :auto))
-    auxNew_Names = [Symbol("Resource"); Symbol("Zone"); [Symbol("t$t") for t in 1:T]]
-    rename!(dfStorage, auxNew_Names)
-    CSV.write(joinpath(path, "storage.csv"), dftranspose(dfStorage, false), header = false)
+    stored_ids = convert(Vector{Int}, vcat(STOR_ALL, HYDRO_RES, FLEX, VS_STOR))
+    df = DataFrame(Resource = resources[stored_ids],
+        Zone = zones[stored_ids])
+    df.AnnualSum = stored * weight
 
-    if setup["OutputFullTimeSeries"] == 1 && setup["TimeDomainReduction"] == 1
-        write_full_time_series_reconstruction(path, setup, dfStorage, "storage")
-        @info("Writing Full Time Series for Storage")
-    end
+    write_temporal_data(df, stored, path, setup, "storage")
 end