+++ CHANGELOG.md
@@ -7,6 +7,35 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.4.2] - 2024-12-23
+
+### Added
+- Fusion plant optional features for thermal plants (#743).
+- Support for reusing the same Gurobi environment for multiple solves when 
+number of concurrent Gurobi uses is limited (#783).
+- Additional long-duration storage constraints to bound state of charge in 
+non-representative periods (#781).
+- New version of `add_similar_to_expression!` to support arrays of `Number`s. (#798)
+- New settings flag `LDSAdditionalConstraints` to provide flexibility in 
+activating new long-duration storage constraints (#781). Can be set in the GenX 
+settings file (PR #801).
+
+### Changed
+- The `charge.csv` and `storage.csv` files now include only resources with 
+charge and storage variables (#760 and #763).
+- Deduplicated docs on optimized scheduled maintenance for thermal resources (#745).
+- Removed the `CapRes_*` columns from `Network.csv` since they were not being used (#784).
+
+### Fixed
+- Add constraint to ensure that electricity charged from the grid cannot exceed 
+the charging capacity of the storage component in VRE_STOR (#770).
+- Update `getproperty` function for vectors of resources to ensure compatibility 
+with Julia v1.11 (#785).
+- Fixed cost calculation in `write_costs.jl` when no resources are present in 
+a zone. (#796)
+- Added `eTotalCMaxCapSlack` to calculation of `cUnmetPolicyPenalty` in 
+`write_costs.jl` (#806).
+
 ## [0.4.1] - 2024-08-20
 
 ### Added